requirements.txt
easy-entrez>=0.3.7
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
streamlit>=1.29.0
plotly>=5.17.0
pandas>=2.0.0
